do sorting on group data level
