+++ app/components/demonstrator-group-data.js
@@ -1,7 +1,4 @@
 import Ember from 'ember';
-import RSVP from 'rsvp';
-import config from '../config/environment';
-import { dateformat } from '../helpers/dateformat';
 
 export default Ember.Component.extend({
   store: Ember.inject.service(),
@@ -11,18 +8,45 @@ export default Ember.Component.extend({
   header: ['Név', 'Neptun', "Csoport", 'Feladattípus', 'Feltöltés ideje', 'Beugró érdemjegy', 'Beadandó érdemjegy', 'Labor érdemjegy', 'Pót'],
   rowIndecies: ['StudentRegistration.User.displayName', 'StudentRegistration.User.neptun', 'CourseCode', 'ExerciseSheet.ExerciseType.shortName', 'firstCorrectableDeliverable.formattedLastSubmittedDate', 'firstEntryTest.grade', 'firstCorrectableDeliverable.grade', 'grade', 'supplementary'],
 
-  eventGroup: Ember.computed('events', function(){
-    return this.get('events').then (function(events){
+  sortedEvents: Ember.computed('events', 'events.[]', function() {
+    return this.get('events').then(function(events) {
+      return events.sort((lhs, rhs) => {
+        const lhsName = lhs.get('StudentRegistration.User.displayName'), rhsName = rhs.get('StudentRegistration.User.displayName');
+        const lhsAttempt = lhs.get('attempt'), rhsAttempt = rhs.get('attempt');
+
+        if (lhsAttempt > rhsAttempt) {
+          return -1;
+        }
+
+        if (lhsAttempt < rhsAttempt) {
+          return 1;
+        }
+
+        if (lhsAttempt === rhsAttempt) {
+          if (lhsName < rhsName) {
+            return -1;
+          }
+          if (lhsName > rhsName) {
+            return 1;
+          }
+        }
+        return 0;
+      })
+    })
+  }),
+
+  eventGroup: Ember.computed('events', function() {
+    return this.get('events').then (function(events) {
       return events.get('firstObject.CourseCode');
     });
   }),
-  eventPlace: Ember.computed('events', function(){
-    return this.get('events').then (function(events){
+  eventPlace: Ember.computed('events', function() {
+    return this.get('events').then (function(events) {
       return events.get('firstObject.location');
     });
   }),
-  eventDate: Ember.computed('events', function(){
-    return this.get('events').then (function(events){
+  eventDate: Ember.computed('events', function() {
+    return this.get('events').then (function(events) {
       var date = events.get('firstObject.date');
       return date;
     });
