+++ app/components/demonstrator-group.js
@@ -21,28 +21,6 @@ export default Ember.Component.extend({
               event.set('formattedDate', dateformat([event.get('date')]));
               return event;
             })
-            .sort((lhs, rhs) => {
-              const lhsName = lhs.get('StudentRegistration.User.displayName'), rhsName = rhs.get('StudentRegistration.User.displayName');
-              const lhsAttempt = lhs.get('attempt'), rhsAttempt = rhs.get('attempt');
-
-              if (lhsAttempt > rhsAttempt) {
-                return -1;
-              }
-
-              if (lhsAttempt < rhsAttempt) {
-                return 1;
-              }
-
-              if (lhsAttempt === rhsAttempt) {
-                if (lhsName < rhsName) {
-                  return -1;
-                }
-                if (lhsName > rhsName) {
-                  return 1;
-                }
-              }
-              return 0;
-            })
           );
         }, err => {
           console.error(err);
